+++ src/cl.cxx
@@ -13,66 +13,12 @@
  *             SPDX-License-Identifier: (Apache-2.0 OR MIT)
  *
  */
-#include "cl.hxx"
+#include "cl.h"
+#include "utils.h"
 
 #include <sstream>
 
-// String helper methods adding cxx20 features to cxx14
-bool startswith(std::string &arg, std::string &match) {
-    size_t matchLen = match.size();
-    if ( matchLen > arg.size() )
-        return false;
-    return arg.compare(0, matchLen, match) == 0;
-}
 
-bool startswith(std::string &arg, const char * match) {
-    return startswith(arg, (std::string)match);
-}
-
-bool endswith(std::string &arg, std::string &match) {
-    size_t matchLen = match.size();
-    if ( matchLen > arg.size() )
-        return false;
-    return arg.compare(arg.size() - matchLen, matchLen, match) == 0;
-}
-
-bool endswith(std::string &arg, char const* match) {
-    return endswith(arg, (std::string)match);
-}
-
-std::string ConvertWideToANSI(const std::wstring &wstr) {
-    int count = WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), wstr.length(), NULL, 0, NULL, NULL);
-    std::string str(count, 0);
-    WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), -1, &str[0], count, NULL, NULL);
-    return str;
-}
-
-std::wstring ConvertAnsiToWide(const std::string &str) {
-    int count = MultiByteToWideChar(CP_ACP, 0, str.c_str(), str.length(), NULL, 0);
-    std::wstring wstr(count, 0);
-    MultiByteToWideChar(CP_ACP, 0, str.c_str(), str.length(), &wstr[0], count);
-    return wstr;
-}
-
-StrList split(std::string s, std::string delim) {
-    size_t pos_start = 0, pos_end;
-    size_t delim_len = delim.length();
-    std::string token;
-    StrList res = StrList();
-
-    while ( (pos_end = s.find(delim, pos_start)) != std::string::npos ) {
-        size_t token_len = pos_end - pos_start;
-        token = s.substr(pos_start, token_len);
-        pos_start = pos_end + delim_len;
-        if (token == delim || token.empty())
-        {
-            continue;
-        }
-        res.push_back(token);
-    }
-    res.push_back(s.substr(pos_start));
-    return res;
-}
 
 std::string getSpackEnv(const char* env) {
     char* envVal = getenv(env);
@@ -346,8 +292,8 @@ void IntelFortranInvocation::loadToolchainDependentSpackVars(SpackEnvState &spac
 }
 
 
-std::unique_ptr<ToolChainInvocation> ToolChainFactory::ParseToolChain(int argc, char const* const * argv) {
-    std::string command(argv[0]);
+std::unique_ptr<ToolChainInvocation> ToolChainFactory::ParseToolChain(char const* const * argv) {
+    std::string command(*argv);
     char const* const* cli(++argv);
     stripPathandExe(command);
     using lang = ToolChainFactory::Language;
@@ -414,22 +360,4 @@ void validate_spack_env() {
 
 void die(std::string &cli ) {
     throw SpackCompilerException(cli);
-}
-
-int main(int argc, char* argv[]) {
-    // Ensure required variables are set
-    // validate_spack_env();
-    // Determine which tool we're trying to run
-    auto c = argv;
-    c++;
-    std::unique_ptr<ToolChainInvocation> tchain(ToolChainFactory::ParseToolChain(argc, argv));
-    // Establish Spack compiler/linker modifications from env
-    SpackEnvState spack = SpackEnvState::loadSpackEnvState();
-    // Apply modifications to toolchain invocation
-    tchain->interpolateSpackEnv(spack);
-    // Execute coolchain invocation
-    tchain->invokeToolchain();
-    // Any errors caused by the run are reported via the
-    // toolchain, if we reach here, we've had success, exit
-    return 0;
 }
