+++ src/winrpath.cxx
@@ -1,77 +1,23 @@
 #pragma once
-#include <sstream>
 #include "winrpath.h"
+#include "utils.h"
+
+#include <sstream>
 
 #define BUFSIZE 4096
 
-using StrLst = std::vector<std::string>;
 using CoffMembers = std::vector<coff_entry>;
 
-// String helper methods adding cxx20 features to cxx14
-bool startswith(std::string &arg, std::string &match) {
-    size_t matchLen = match.size();
-    if ( matchLen > arg.size() )
-        return false;
-    return arg.compare(0, matchLen, match) == 0;
-}
-
-bool startswith(std::string &arg, const char * match) {
-    return startswith(arg, (std::string)match);
-}
-
-bool endswith(std::string &arg, char const* match) {
-    return endswith(arg, (std::string)match);
-}
-
-bool endswith(std::string &arg, std::string &match) {
-    size_t matchLen = match.size();
-    if ( matchLen > arg.size() )
-        return false;
-    return arg.compare(arg.size() - matchLen, matchLen, match) == 0;
-}
-
-std::string ConvertWideToANSI(const std::wstring &wstr) {
-    int count = WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), wstr.length(), NULL, 0, NULL, NULL);
-    std::string str(count, 0);
-    WideCharToMultiByte(CP_ACP, 0, wstr.c_str(), -1, &str[0], count, NULL, NULL);
-    return str;
-}
-
-std::wstring ConvertAnsiToWide(const std::string &str) {
-    int count = MultiByteToWideChar(CP_ACP, 0, str.c_str(), str.length(), NULL, 0);
-    std::wstring wstr(count, 0);
-    MultiByteToWideChar(CP_ACP, 0, str.c_str(), str.length(), &wstr[0], count);
-    return wstr;
-}
-
-StrLst split(std::string s, std::string delim) {
-    size_t pos = 0, start = 0;
-    std::string token;
-    StrLst res = StrLst();
-    for (;;) {
-        pos = s.find(delim, start);
-        token = s.substr(start, pos-start);
-        start = pos+1;
-        if (token == delim || token.empty()) {
-            continue;
-        }
-        res.push_back(token);
-        if (pos == std::string::npos) {
-            break;
-        }
-    }
-    return res;
-}
 
 LinkerInvocation::LinkerInvocation(std::string linkLine): line(linkLine) {
     this->is_exe = true;
-    this->libs = StrLst();
-    this->tokens = StrLst();
+    this->libs = StrList();
+    this->tokens = StrList();
 }
 
 void LinkerInvocation::parse() {
 
-    StrLst tokenized_line = split(this->line, " ");
+    StrList tokenized_line = split(this->line, " ");
     this->tokens = tokenized_line;
     for (auto token = this->tokens.begin(); token != this->tokens.end(); ++token) {
         if (endswith(*token, ".lib")) {
