+++ test/calc.cpp
@@ -1,0 +1,6 @@
+#include "calc.h"
+
+int add(int &a, int &b)
+{
+    return a + b;
+}
